use EndCallHandler in App
