+++ src/components/index.js
@@ -10,6 +10,7 @@ export * from './ArticlePreview'
 export * from './ArticleToc'
 export * from './ConfirmDialog'
 export * from './Consent'
+export * from './EndCallHandler'
 export * from './Error'
 export * from './Feed'
 export * from './Feedback'
